Update patch set 20
Patch Set 20: Code-Review+2

I'm +2 merging this as the build is now working in CI.  I agree with Randy that there is some cleanup to do, but we can address that in another patch.  For now it is important to get daily builds running in CI so we can get artifacts to test deploy with.

Patch-set: 20
Label: Code-Review=+2