Update patch set 1
Patch Set 1: Code-Review-1

Instead of doing all that file doctoring I think it would make more sense to generate those configs in the build like have done with the non-ha stuff. I had started a patch a couple days ago for this:
https://github.com/radez/apex/tree/HA_support

see what you think of that and maybe you could pull that and continue work on that. In my tests the deploy started  but failed and I didn't get a chance to debug the failure

Patch-set: 1