Update patch set 1
Patch Set 1: -Code-Review

I didn't really expect us needing to scale virtually dynamically because of the massive amount of resources needed to to a virtual HA deployment with minimal hardware, not sure the LF hardware will even be able to do it as is. An this won't impact baremetal. My thought was to keep deploy as lean as possible.

Patch-set: 1
Reviewer: Gerrit User 235 <235@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=0