Update patch set 2
Patch Set 2: Code-Review-1

I'm not sure this is the best fix.  The problem I saw was the virsh machines were 2, but instack was trying to introspect 5 nodes (which makes me think they are still in the instackenv json).  I think its better to just remove the extra nodes so that unused virsh machines aren't created/introspected for a non-ha deploy.

Patch-set: 2
Label: Code-Review=-1