Update patch set 2
Patch Set 2:

Right, I did it this was because it was simple.
It still defines the 5 nodes that the instackenv file expects to see. Then no matter which ones get choosen they are there to be booted.

My intent was not to bloat the deployment scripts with trying to handle cleaning up the instackenv file when defining and not booting some vms that will be automatically torn down when it's all cleaned up is a cheap operation.

Patch-set: 2
Reviewer: Gerrit User 235 <235@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0