Update patch set 6
Patch Set 6: Code-Review-1

This is a good first patch to get external working, however we need the network isolation to be not only external but as OPNFV requires.  We need to split admin, private and storage as well.  We should define a ServiceNetMap to isolate the services to the logical networks.

Patch-set: 6
Label: Code-Review=-1