Parses yaml files into deploy + other fixes
This patch allows deploy_settings.yaml and inventory settings to be
passed in via cmdline.  The deploy settings are overridden by cmdline
arguments.  For example, if a user specifies no_ha in cmdline it will
override what is in deploy_settings.yaml.  The inventory file provides
yaml input for the baremetal topology.

Fixes include:
 - New arguments for passing inventory and deploy_settings
 - "global_params" in deploy_settings are set as globals in script
(unless already set via cmdline)
 - "deploy_options" are loaded into an array, and checked for deployment
options (such as which SDN controller to use)
 - inventory file is parsed (for baremetal) into correct instack json
output.  The output is then written to a temp instackjson file and used
for deployment.
 - renaming of copy_materials to copy_materials_to_instack
 - spelling fixes
 - cmdline argument checking with errors and warnings

Note this patch does not yet tamper with setting network-environment
settings.

JIRA: APEX-53

Change-Id: I9dce197e23563480c584650ba967898312eab4e7
Signed-off-by: Tim Rozet <trozet@redhat.com>