Update patch set 1
Patch Set 1:

Viktor awesome work!  I didn't know this was an option to the overcloud_deploy, but we needed a way to dictate which nodes were control and compute.  From looking at the failure it looks like capabilities property is defined twice.  Just need to fix it to be part of the same property and then I think it will work.

Patch-set: 1