Update patch set 1
Patch Set 1:

Agreed, nice work. Thanks for the patch. Looking forward to seeing it pass verify. One note for you, when you push a new review try and remember to rebase on origin/master before you push it to ensure you're on top of the latest and avoid conflicts.

Patch-set: 1