Update patch set 2
Patch Set 2:

(2 comments)

I like the patch.  Just a few suggestions about making it easier to find.

Also the file is being overwritten with each deploy.  What about archiving the files and adding some time and date info either inside the file or to the file name?

If you do archive them, it may be better to put them in a dedicated subdirectory.

Patch-set: 2