Update patch set 2
Patch Set 2:

I think the idea is to expand this in the future and make it more user friendly to redeploy without having to tear down the undercloud.

This commit is just to help up debug using the command that was initially used to deploy with.

Patch-set: 2
Reviewer: Gerrit User 235 <235@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0