Update patch set 1
Patch Set 1: Code-Review-1

(2 comments)

Good stuff in here Bob,
The external network was just added in the past day or two, you may not have gotten a build with it yet.

The forwarding on the external network would be great for all virtual deployments.

Finally, I think the sh script should be named something with onos in it like onos_gw_mac_update so that it is evident that this is an onos specific sh script within the apex code base.

Thanks for keeping on top of things and contributing. Your efforts are appreciated!

Patch-set: 1
Label: Code-Review=-1