Update patch set 1
Patch Set 1:

After talking to Ash, onos also supports baremetal just fine.

Now that rpm split is fixed, and onos is included in the build.  This is a good chance to test out this patch against a verify onos jjb (which is pending 2 patches).  I will update this patchset and push it once that verify job exists.

Patch-set: 1