Update patch set 4
Patch Set 4: Code-Review-1

(3 comments)

Nice work Michael.  There are a few issues inline.  One of them is my mistake, hope you dont mind fixing that.  This is a good opportunity to add ODL L3 to the verify jobs.  Will do that shortly so that we capture a verify on it in your patch set.

Patch-set: 4
Label: Code-Review=-1