Allows ONOS to use compute nodes with br-ex no IP
ONOS needs the capability to have br-ex on compute nodes with no IP for
L3 functionality to work correctly on baremetal.  This patch allows that
capability, and makes it default behavior (sdn_l3 true).  If desired, a
user can set sdn_l3 to false, and compute nodes public interface will be
a regular interface with no IP for ONOS deployments.

JIRA: APEX-64

Change-Id: I578b946edbddf7fc9c29363c78fec2c3ada48cf6
Signed-off-by: Tim Rozet <trozet@redhat.com>