Updating directory structure to add files to the central docs.
Moved the introduction and baremetal installation instructions to /configguide.
The files baremetal and introduction will be pulled into the central
config guide document for release.  (A little local file name juggling
was required to accommodate the toolchain.)
Once merged this should be pulled into the configguide.  Will work on a 
patch for that now.

Change-Id: Ie270154d0c7ac418b693ac7cf6e1444a33023266
Signed-off-by: ChristopherPrice <christopher.price@ericsson.com>