Update patch set 1
Patch Set 1: Code-Review-2

After thinking about this some more, I agree with Tapio.  The correct behavior on master can be to grab latest, but we need to freeze the Be artifact and puppet-odl we pull.

We will need a patch to modify how we pull the ODL Be artifact on stable/b, so that it is a frozen and specific version.

Patch-set: 1
Reviewer: Gerrit User 334 <334@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=-2