Update patch set 1
Patch Set 1:

> Tapio brings up a good point.  We are switching to a moving target
 > here, and we are trying to freeze stable/b.  Not sure what to do
 > here.

 > Tapio brings up a good point.  We are switching to a moving target
 > here, and we are trying to freeze stable/b.  Not sure what to do
 > here.

The problem is that ODL Be isn't out yet, so it's a bit hard to provide a tag that installs it already.

One option is to pin to a version range of tags, which we could do via some git/shell logic.

Another would be for me to try to cut a puppet-opendayligt release that would work for Be once Be is released. You could freeze to that tag on stable/b and use master for integrating Be ERPs pre-release. Eventually the ERP consumed by master will be blessed as official Be by the ODL TSC, we'll move it to the nfv7-opendaylight-4-release CBS tag and the puppet-opendaylight version you're tagged on would start pulling it (so master and stable/be would point at the same version via different methods).

Patch-set: 1