Update patch set 4
Patch Set 4:

because you only need to reload the daemon if you are going to start it right way. The daemon loads the file when the overcloud nodes boot for the first time. All that needs to happen now is enable it so that when the daemon loads on boot on the overcloud nodes the service is marked to start.

Patch-set: 4
Label: Verified=0