Update patch set 1
Patch Set 1:

> (1 comment)
 > 
 > Jamo this should take care of adding an VIP and configuring HA
 > Proxy.  It will load balance between all 3 ODLs for neutron
 > requests.  If we need to isolate requests to only a single ODL node
 > at a time (like stateful OS services with pacemaker), I'm not sure
 > how to do that.  Dan, do you know?  Will look into it.

I've never used haproxy specifically, and not sure if we can configure
it easily or not.  I think if our LB algo is something like "source" [0]
we can guarantee that we are always going to hit the same ODL instance
with every request.  If the node we are hitting fails, then it will
pick a new instance but it will stick with that instance going 
forward.

[0] http://www.haproxy.org/download/1.5/doc/configuration.txt

I heard during an ODL+OpenStack session that there is worry about
race conditions when the requests are spread over multiple 
connections.  It could be the issue in networking-odl I've heard
before, or even with ODL.  Or maybe no issue.  Right now it's just
hearsay I guess.

Patch-set: 1