Update patch set 6
Patch Set 6:

The issue in the latest failure was ODL was attempting to bind to 0.0.0.0 when HA proxy has already bound to it's VIP.  Have made changes to puppet-opendaylight to allow ODL to bind to a specific IP address in my github.  Next patch set will include that fix.

Patch-set: 6