Update patch set 1
Patch Set 1:

Merged stable/mitaka into stable/colorado.  The issue is the controller now looks for it's image based on the Image variable in controller.yaml:
https://github.com/trozet/opnfv-tht/blob/57851901c557ec388d7efbde8f1ef9b7c4ee3bae/puppet/controller.yaml#L279

Which is inherited from controllerImage here:
https://github.com/trozet/opnfv-tht/blob/stable/colorado/overcloud.yaml#L942

I think we need to update our env to point this var to be our overcloud-full image.

Patch-set: 1