Update patch set 1
Patch Set 1: Code-Review-1

I'd prefer if the new file was named something relative to the project. opnvf-environment or apex-environment. That way there's an indication that it's a value specific to our deployment.

I'd also prefer that it be dropped onto the undercloud image at build time. There's no need for that to be uploaded at deploy time, I think we should push as many boiler plate tasks as we can to the build.

Patch-set: 1
Label: Code-Review=-1