Update patch set 1
Patch Set 1:

NOSDN is already in and pretty stable. I think we're at the point that we need to iterate on this to stabilize it and the question is "do we do that on the verify and block everything else or in the daily to keep verifys going?"

I'm of the opinion that this is fundamental to the release being viable and that we should block on it. Other things in flight can and probably should wait for this to verify properly with ODL and ONOS verify tasks.

Patch-set: 1