Update patch set 1
Patch Set 1:

we do pull the latest PR and use it on the undercloud when opnfv-tht-pr is specified in the commit msg.  We should use a dummy log file to record PR revisions when we have no other Apex code to modify in a commit.

Patch-set: 1
Reviewer: Gerrit User 334 <334@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0