Add custom domain file for virtual deploy
Upstream packaging for tripleo-incubator seems to be lagging
significantly, and in addition we will likely want to add a myriad
of custom features like numa support that may not be suitable for
upstream, so copy our own domain.xml into the tripleo templates
directory when deploying.

This may be improved in future if we can land a patch that
allows a custom template directory to be specified upstream, and
have it packaged and released.

Change-Id: I7e7e7f695d9473545014761e9f8986616bc2cd19
JIRA: APEX-119
JIRA: APEX-106
JIRA: APEX-127
Signed-off-by: Michael Chapman <woppin@gmail.com>
Signed-off-by: Tim Rozet <trozet@redhat.com>