Update patch set 5
Patch Set 5: Code-Review-1

We want to do nosdn first, to keep it simple.  Also the naming of the scenario file needs to be according to: https://wiki.opnfv.org/display/SWREL/Colorado+scenario+inventory+and+dependencies

os-nosdn-ovs-ha
os-nosdn-ovs-noha

ovs as a feature here implies dpdk.  I know it would be better to use dpdk in the name to be specific.  I'll bring it up in the call, but for now please use the standard naming scheme and add ha as another scenario.

Patch-set: 5
Label: Code-Review=-1