Remove swift endpoint and service
Swift is not functional, and ceph is used instead for glance, nova and cinder,
so the endpoint should not be present in the overcloud. The endpoint is
removed via CLI post-install since there is no parameter for overcloud.yaml
to disable endpoint creation, and the mechanism for endpoint creation will
likely change within the next release as it is moved from heat to puppet.

JIRA: APEX-146

Change-Id: If91b9d2747d5eed89c569ed9a34ab84370c3d65f
Signed-off-by: Michael Chapman <woppin@gmail.com>