Update patch set 3
Patch Set 3: Code-Review-1

I think this is the right direction (unit tests with coverage), but I would prefer to see this type of test done during jenkins verify as its own job and not part of the build.  Ideally we would want this to run before anything else.

Patch-set: 3
Label: Code-Review=-1