Allows nic mapping to set hieradata variables
Currently os-net-config figures out the mapping of logical nic names,
such as "nic1" to a physical nic on the overcloud node.  This mapping is
not exposed to THT.  This patch allows the mapping to be placed in a
hieradata file that can be access by THT during the deployment.

JIRA: APEX-119

opnfv-tht-pr: 28

Change-Id: I7ec631fe2a04ce6193b72b2e56eb34a24a4507e1
Signed-off-by: Tim Rozet <trozet@redhat.com>