Fixes nic template rendering
Issue where in default network_settings, api_network is not specified.
This means that the settings do not have api_network in its output, but
the nic template checks to see if api_network vlan is native in its
logic.  This patch assumes a vlan will be numeric. A blank value or a
value of 'native' will not be numeric therefore default is expanded to
allow blank values as non-vlan networks.

Also fixes controller to always have br-ex.

Change-Id: Ibebb231901e653127175aef194eec708e8507da2
Signed-off-by: Tim Rozet <trozet@redhat.com>
Signed-off-by: Dan Radez <dradez@redhat.com>
Signed-off-by: Tim Rozet <trozet@redhat.com>