Fixes nic template rendering
Issue where in default network_settings, api_network is not specified.
This means that the settings do not have api_network in its output, but
the nic template checks to see if api_network vlan is native in its
logic.  This patch defaults all possible networks to native if they are
missing in settings and fixes the template rendering.

Change-Id: Ibebb231901e653127175aef194eec708e8507da2
Signed-off-by: Tim Rozet <trozet@redhat.com>