Finishes dpdk post install
After dpdk finishes install, br-phy needs to be brought up and ovs-agent
needs to be restarted.  This cannot happen in tripleO because
os-collect-config will try to ping from dpdk interfaces to the
controller to validate network connectivity.  This doesn't work with
dpdk, so we leave br-phy down until end of deployment.

JIRA: APEX-119

Change-Id: I3d36d02b773fcbb3c5527b85fd00a2fa1b3cc25d
Signed-off-by: Tim Rozet <trozet@redhat.com>