Update patch set 3
Patch Set 3:

(1 comment)

Can you add some context to this commit message/JIRA?  What is actually being accomplished here?  VSPERF is added to the compute nodes, but how is that used?  I thought from our discussion with Mars the performance tool would be run on a compute instance, not the compute node?

Patch-set: 3