Update patch set 4
Patch Set 4: -Code-Review

Oops, didn't see Mars responded.  So then how does vsperf work if it is running on a compute node?  Can we move this conversation to JIRA and please provide some background to the JIRA on how this feature works.  Otherwise, I don't know how to review this.

Patch-set: 4
Label: Code-Review=0