Allows specifying nic order for overcloud nodes in network settings
Currently there is no way to specify logically or physically the nic
order to be used on overcloud nodes.  We always hardcode to use nic1 for
admin network, nic2 for private, etc.  This patch allows a user to not
only decide which logical nics to use for which network, but also
specify physical interface names if they need to.

This is done on a per role basis, due to tripleO limitation.  So a user
is able to specify nic order/names for compute and controller roles
separately.

If a user specifies nic order, they must specify it for all networks
other than admin network.  We assume if admin network is unspecified it
uses "nic1", so that name is reserved in this case.  A user is also
allowed to specify a mixture of logical and physical names, for example
"nic2" and "eth3" on another network.

JIRA: APEX-151

Change-Id: Ie9d4abb463cf8f8788913cb4bcf9486830adc449
Signed-off-by: Tim Rozet <trozet@redhat.com>