Update patch set 10
Patch Set 10:

Also, can you add the deploy config files and change their default to true for tacker.  We want it enabled by default, but the user can selectively turn it off (in case they want to use a different orchestrator).  Can you also add a check after hte deployment that if tacker was enabled as a deploy option, just do an openstack-status and make sure it is started?

Patch-set: 10