Update patch set 2
Patch Set 2:

After talking with Feng, I think we can push this in after merge conflict is fixed.  The goal is to get this working yesterday, and Feng's patch will help us do that.  We will work on fixing/changing the puppet-fdio in your repo on the side, so we can make changes there without worrying about blocking progress in Apex.  Once it is ready we can just do a full swap for that module.

Patch-set: 2
Label: Verified=0