Update patch set 3
Patch Set 3:

I'm using the VM_COMPUTES var to hold what the use passes, we have to be able to know if a value has been passed or not by the time it gets to the calculation in the virtual setup.
VM_COMPUTES=0 indicates that we did not get a value from the end user and that the defaults should be used.

Patch-set: 3
Label: Verified=0