converiting the deploy settings obj to a dict
The deploy settings values are embedded in a dictionary inside a
generic object. This patch makes the deploy settings object
a dictionary so the values can be accessed directly without
having to unnecessarily drill down through an empty object
to the dict that holds the content intended to be managed by
the deploy settings object.

- adding tests to cover DeploySettings 100%

Change-Id: I4ba625cd7b51cfb6c1f91c74f1d332d1e3dd9a8e
Signed-off-by: Dan Radez <dradez@redhat.com>