Fixes control/compute node detection
Previous behavior was detecting the total number of nodes in the
inventory file and defining number of compute nodes as total nodes,
minus 1 or 3 control nodes (depending on HA flag).  This patch changes
the behavior to look explicitly for number of compute/control
nodes in the inventory file.

JIRA: APEX-244

Change-Id: I0191bb88e7dc5bebb5f7bfed39661f8047dbd476
Signed-off-by: Tim Rozet <trozet@redhat.com>