Update patch set 1
Patch Set 1:

> I'm not sure we should get into carrying puppet code like this in
 > the apex repo.
 > 
 > It's a 3rd way of doing something we already have two ways of doing
 > (forking or .patch file)

We already do this today with L3 agent in build/neutron, but yeah that's fine if we want to stick with forking/patch. I can change this to a patch file, it just seemed more clear to have the files since the change is just to add additional files to puppet-neutron.

Patch-set: 1