Apply vlan_strip_offload config only when using vlan for fdio
We currently apply vlan_strip_offload cnofig to VPP interface for
all enic interfaces, however, this is really only needed when
vlan tagged traffic is being passed on the interface. This patch
adds a vlan config option to puppet-fdio and opnfv-tht so that
we only enable it when needed. Right now the only scenario this
is needed is nosdn-fdio, since the only supported mode for tenant
networking is VLAN.

Dependent on https://github.com/radez/puppet-fdio/pull/11

JIRA: APEX-285

opnfv-tht-pr: 87

Change-Id: I40c444896e036eaa87506cf372bedd4b8a621d00
Signed-off-by: Feng Pan <fpan@redhat.com>