Fix tenant_nic loopup error for nosdn fdio scenario
Currently we assume tenant_nic is a virtual interface name (like nic1),
so we perform an extra hiera lookup, this is not needed if the user
specifies physical nic name in network_settings file. This patch defaults
the hiera lookup of the nic interface lookup to be what the user specifies
if a hiera value can't be found.

JIRA: APEX-293

opnfv-tht-pr: 88

Change-Id: I5f6d3addbaf20c4dfec532b1292ddb9424552a1f
Signed-off-by: Feng Pan <fpan@redhat.com>