Update patch set 2
Patch Set 2:

I needed to merge this for my other patch that's updating the caching. The file was renamed to dev_dpe_check because we have both building and deployment dependencies in that file, It's in the ci dir which I think is a bit misleading.

Let's continue discussion on where to put this file. I think we can clean it up more so that it's clearer. For now it atleast cleans up the perpetual downloading of the files that are already installed.

Patch-set: 2