Enabling alternalte method to get remote md5
The onos tar is an example of the remote server not supplying
a .md5 file with the md5 in it. It is instead posted on an html
page that can be scraped to get the value. This patch adds the ability
to pass that value into the caching functions so can keep onos
cached and not redownload it over and over again

JIRA: APEX-325

Change-Id: I88be72d4fd6aa7245b02b9ef0f2b7e6eefff3dac
Signed-off-by: Dan Radez <dradez@redhat.com>