Update patch set 15
Patch Set 15:

It's a build script used by CI and for individual's development, So both cases apply.

Our developers integration guide instructs developers to use this method to build so we have to treat the case that someone builds in place multiple times during their development.

We've been discussing cleaning up the build tree, this is will help to treat a polluted build tree.

Patch-set: 15