Update patch set 15
Patch Set 15:

With the built artifacts, eg. the tacker RPM is built from a branch, but has a hardcoded file name and version. I think if the upstream branch is updated, we wouldn't be rebuilding the RPM, because the file name doesn't change, right? So to cache those files and still be able to pull in upstream changes, we would need to have commit count or sha incorporated into the RPM name.

Patch-set: 15
Label: Verified=0