Update patch set 16
Patch Set 16:

We have two options there, we could use the rpm spec we carry to update the release number when we want a new version and that would change the file name, or a hash could be built into the RPM release number to trigger a build.

We also have the option of storing a hash if we want to tie ourselves to a specific upstream commit or HEAD if we just want the latest.

Patch-set: 16
Label: Verified=0