Update patch set 3
Patch Set 3: Code-Review-2

I think the concept here is ok, but we shouldn't add this file without having the underlying support for what it's doing added into the code base.

Adding this file gives the impression that if someone fills it out and uses it that it will work properly. In reality the items that are added to this file will be ignored and the deploy will fail.

We need to work through the configuration that needs to happen on the under cloud to support the configuration proposed here. If you need some help with that please start an email thread or discussion in the JIRA and we can start to identify the work items that need to be done.

Patch-set: 3
Reviewer: Gerrit User 235 <235@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=-2