Adds declaring disk device to use on overcloud nodes
Now in inventory file a user can declare 'disk_device' which will allow
the user to specify which disk on their overcloud server to use for
installation.  The variable can be a comma delimited list, which will
search in order post-introspection for the first device on the node that
exists.

Currently defining a per node disk is not supported by OSCLI, so
although defined that way in Apex inventory, only the last definition
will be used for all nodes.

Other changes include:
 - Various fixes in inventory parsing
 - Makes bash writing a common function
 - Introspection now enabled for baremetal deployments

JIRA: APEX-296

Change-Id: I330d91eb17408ccfceb7a99c25edbae5ce6d848d
Signed-off-by: Tim Rozet <trozet@redhat.com>