Update patch set 3
Patch Set 3: Code-Review-1

I'm not sure about calling it easy install. First because this is running a deployment of apex rather than merely an installation of the packages, and secondly because that term is closely associated with the python command. I would suggest maybe calling it 'easy deploy' or 'simple deploy' or something along those lines.

Patch-set: 3
Label: Code-Review=-1