Update patch set 1
Patch Set 1: Code-Review-2

I'm not sure if this is going to work. Since there is no heat parameter, I just provided it in extra config, however the same variable is already provided in THT, so not sure how that works.  Will THT merge it and use the one I provide, or will it create 2 hiera entries and it will be random which one it picks?  Need to test and see.

Patch-set: 1