Update patch set 2
Patch Set 2:

I don't think the upstream one will work,
it depends on macros that seem to be specific to koji doing the builds.

The point of getting rid of this file is that RPM should mamage the deps and not let python try and handle the deps.

IT's not really a hack I think it's kinda the right way to do it. It's the way upstream does it.

Patch-set: 2
Reviewer: Gerrit User 235 <235@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0