Update patch set 4
Patch Set 4: Code-Review-1

(2 comments)

I'm not deeply familiar with yardstick, but if it's possible to install it within the undercloud and run from there, I think that would be better as we avoid any issues surrounding jumphost reuse in the case of bugs or failed cleans. This would also make it pretty easy for us to add the gui and API endpoints in a neat way (via undercloud customisation).

Patch-set: 4
Label: Code-Review=-1