Update patch set 18
Patch Set 18:

Hey guys, do we really want to build quagga on every build of apex? We are not the CI from quagga or? I think a better solution is to build it once and then store the rpm somewhere.

We can save the build script in apex so that we have it easier to build the next rpms next time when we need an update from quagga.

In addition I saw that the thrift repo is VERY unstable. I get http errors VERY often so that our build system would as well fail due to that and when people build on their desktop machines it will fail as well from time to time. They will always come to us and ask us what is wrong.

Patch-set: 18
Label: Verified=0