Update patch set 18
Patch Set 18:

We are locked in thrift and capnproto to specific git tags.  I think those should build fine.  I think we want to build zrpc and quagga each time so that we pick up any fixes we need.  If we find that the build process is taking too long, or there are issues with repos going down, we can look at storing packages somewhere and downloading rather than building.

Patch-set: 18
Label: Code-Review=0