Increases open files limit for mariadb
There is currently an issue where the max open files limit is hit with
MariaDB in noha deployments, because it is defaulted to 1024 by system
limits. In HA deployments the limit is bumped to 16384. This patch
introduces a flag to be able to increase the limit to 16384 for noHA
deployments.

opnfv-tht-pr: 115
opnfv-puppet-tripleo: 17

JIRA: APEX-345

Change-Id: I786517c5fb663332e782500df964a43d3d4538ff
Signed-off-by: Tim Rozet <trozet@redhat.com>