Removes building quagga/zrpc
Since building this takes 10 minutes or so and it has intermittent
failures, we are moving it to artifacts.  If we need timely changes we
can implement a job to build it, using the apex makefile.

JIRA: APEX-403

Change-Id: Ie26106062143c968492dce52149459a8dc7a47ba
Signed-off-by: Tim Rozet <trozet@redhat.com>
(cherry picked from commit d5527ec6ff391ddd3f872b7f8a7bc3f1f2075381)