Update patch set 2
Patch Set 2:

Note this was a regression when we submitted a patch to pin all ODL versions for vpp due to repo incompatibility issues.  The ODL version we had decided to use for fdio was no longer being used, and instead latest carbon was being used.  This patch fixes that.

Patch-set: 2