Update patch set 24
Patch Set 24: Code-Review-1

It doesn't look like controllers_ip actually configures the IP ONOS will run on:
https://github.com/bobzhouHW/puppet-onos/blob/master/manifests/service.pp#L26

The cluster code is commented out.  How does it form a cluster?

Also, it looks like each ovs is only configured to connect to 1 ONOS service.  Is this intentional?  Is HA not supported with ONOS?

In your ml2::onos in opnfv-puppet-tripleo, you should point onos_ip to be onos_api_vip, so that it will be load balanced to each ONOS instance.

Patch-set: 24
Label: Code-Review=-1