Install missing packages on undercloud image
Currently, tripleo images are not synchronized to CentOS CDN, where
we download our images. The bug is tracked at:
https://bugs.centos.org/view.php?id=12994

This causes our undercloud image to be out of date and missing
packages which are required during undercloud install. This causes
puppet to perform yum install of those packages during undercloud
deployment, which will fail in cases where internet is not accessible
during deployment.

While we wait for resolution of CentOS CDN issue, to work around
this issue, we need to add those packages in undercloud image.

Change-Id: Ic5a2cfddda544c2d7b6b5bcb145ade8276fd411b
Signed-off-by: Feng Pan <fpan@redhat.com>