Install docker on undercloud image
Currently, tripleo images are not synchronized to CentOS CDN, where
we download our images. The bug is tracked at:
https://bugs.centos.org/view.php?id=12994

This causes our undercloud image to be out of date and missing
docker package which is required during undercloud install, as
docker_registry is enabled by default. This causes puppet to
perform yum install of docker during undercloud deployment, which
will fail in cases where internet is not accessible during deployment.

While we wait for resolution of CentOS CDN issue, to work around
the docker issue, we can either install docker in the undercloud
image or disable docker_registry.

This patch installs docker in the undercloud image, as docker is
also needed for cases when Dovetail is enabled, for example.

Change-Id: Ic5a2cfddda544c2d7b6b5bcb145ade8276fd411b
Signed-off-by: Feng Pan <fpan@redhat.com>