Update patch set 3
Patch Set 3:

(2 comments)

We have to wait to upload the new quagga.tar.gz version.

I think that we could replace the initial and faulty quagga.tar.gz file, instead of uploading a version 3 file.
In such a case we have to update the file name in this patch, too.

Patch-set: 3
Label: Verified=0