Moves RDO images URI to working server
CBS has been broken and the overcloud and undercloud artifacts are not
getting updated there.  The new URI is a workaround because it is a
single file server acting as a backup until the CBS repo is fixed.

JIRA: APEX-458

Change-Id: I743747015d69b17d75aa7b231bc07c935788ea20
Signed-off-by: Tim Rozet <trozet@redhat.com>
(cherry picked from commit ef3dc1ce0323fa0881e416cd4b9028fb4250b719)