Update patch set 9
Patch Set 9:

For the quickstart bits I've just added stubs, then the quickstart patch depends on this one and does the actual implementation, but this way you can structure the main in the way you want. I had a similar rewrite in the quickstart patch so this gets rid of that and moves the pieces I needed into either this patch or into lib/python for use by this.

Patch-set: 9
Label: Verified=0