Update patch set 2
Patch Set 2: Code-Review-1

(5 comments)

Nice job John.  Should the barometer deploy setting be enabled by default on all scenarios, or is this something we are going to have a specific scenario for?  Either way I would like to see it the code path hit in CI.  So please add/update a deploy settings file for it.

Patch-set: 2
Label: Code-Review=-1