Update patch set 14
Patch Set 14: -Code-Review

The failure is due to the controller nodes running out of RAM.  I don't think it is related to this patch as the barometer stuff is only running on compute node.  I think the solution is to bump the minimum ram for controllers to 10GB for nosdn and 12GB for ODL (previously 8, 10).  We can solve that in a separate patch and run the noha gate on this patch which should use less RAM.

Patch-set: 14