Sets baremetal workers to 12
There is an issue where in Newton the number of workers (sub processes)
opened per openstack service scales with the number of logical cores on
the system.  On servers with large amount of processors this causes the
cloud to fail after some time due to running out of RAM.  This patch
will limit the number of worker processes to 12, which is the limit in
Ocata and onwards.

JIRA: APEX-410

Change-Id: I28c4762830c19cd758a0729017f2d4a9ebe372b4
Signed-off-by: Tim Rozet <trozet@redhat.com>
Signed-off-by: Feng Pan <fpan@redhat.com>