Update patch set 4
Patch Set 4: Code-Review-1

(1 comment)

Need a unit test case for testing a negative return code and ensuring exception is raised.

In the log when we fetch or copy files, it prints out the entire file as bunch of unreadable string.  Can we try lowering the log level from -vvv to -vv and see if that is a better level for us?

Patch-set: 4
Label: Code-Review=-1