Update patch set 1
Patch Set 1:

see what you think about this patch, it opens a udp port up during deploy.py and anything we want to output live we can netcat back to the port. I haven't looked into it but it may be better to run it through an ssh port forward instead of the wide open udp port. This seems to work so far, I've been using it in my lab. Interested to get some feedback on it.

Patch-set: 1