Adds ability to deploy from upstream openstack
Note this only will currently work for nosdn.  Ability to add ODL should
be added in separate patch.  To deploy with upstream openstack branch,
use new deploy setting 'os_version'.  A default scenario file for nosdn
with pike has been included in this patch.  If 'os_version' is a version
other than the default version for this OPNFV release, then upstream is
used.

In order to use upstream with the current OS version use '--upstream'
argument to the deploy command, to force an upstream deployment.

Change-Id: Ic07e308827b449637b4e86cdd086434e4de2fb69
Signed-off-by: Tim Rozet <trozet@redhat.com>