Update patch set 4
Patch Set 4:

> (1 comment)
 > 
 > Is barometer deployed in the os-odl-nofeatre-ha job?
 > 
 > Returning node parameter for lf-virtual2
 > Finished Build : 236 Suite: healthcheck Scenario: os-odl-nofeature-ha
 > of Job : functest-apex-virtual-suite-master with status : FAILURE

No barometer is not included in this.  This is failing because somehow the port to download a functest image in LF lab has been blocked again.  Will send an email to LF and try to get it fixed.  This wont stop us from merging your patch.  We will run gate and validate it deploys whent eh patch is ready.

Patch-set: 4
Label: Verified=0