Fix broken build dependency
libguestfs-tools is failing to install because libvirt-client is
installed and there is a package conflict.  Removing libvirt-client from
image before installing libguestfs-tools.

NSH build is failing in overcloud because the CentOS repos updated with
newer kernel source and headers, and we are still using older kernel.
This patch adds the old kernel source and headers.  Upgrading to new
kernel causes NSH OVS to fail when compiling.

Change-Id: Ifd500ad138116cfecafb2268ccb580bd44d5efd5
Signed-off-by: Tim Rozet <trozet@redhat.com>