Update patch set 8
Patch Set 8:

Looks like is finished barometer but ran into trouble with ovs.
I'm not sure what will happened when it builds and runs the bar scenario. The collectd config needs to be changed to match the new collectd-openstack-plugin. I don't know if a failure to run the plugin will cause the scenario to fail.

Patch-set: 8
Label: Verified=0