Fixes Tacker auth and fernet keys dir
Tacker authtoken was missing password which is now added in the tht
patch.  Tacker also needs 'etc/tacker/vim/fernet_keys' in order to be
able to create VIMs.  The real fix to this missing directory is to the
rpm itself:

https://review.rdoproject.org/r/#/c/9725/

However in this patch we ensure it is created in Ansible because
RDO images are frozen in for now in Apex so we would not pick up new
RDO packages.

apex-tripleo-heat-templates: I2b347cbc4595e6651b0d4be032cb862fde72e15f

Change-Id: I76959cbfd581b9fec69749ccbcbcb39dc959a883
Signed-off-by: Tim Rozet <trozet@redhat.com>