Applies Tacker upstream patches and fixes fernet dir perms
We need a couple patches to Tacker upstream that we will not get now
that our images are frozen in RDO.  We should remove them on upgrade to
Pike.  Also this patch fixes the permissions on the fernet keys dir so
that the Tacker group can write.

Change-Id: I6468356f1d16d3b8ab58d46f653431e29960e190
Signed-off-by: Tim Rozet <trozet@redhat.com>
(cherry picked from commit 91a5fa414d46b5b2215bf823166babd034e92503)