Fixes nova migration
Nova migration was failing in rally sanity test.  Looking at the keys
nova is configured with, the newlines in the private key were being
converted to spaces in hieradata, thus making the key invalid.  This
patch corrects the yaml so it is multiline data.

Also, corrects the parameter being set to the right THT parameter and
adds sshd service to the control/compute roles which is required for
migration to work.

JIRA: APEX-525

Change-Id: Ic57db5c0b27df2e1f584f3c1a379e067291d892d
Signed-off-by: Tim Rozet <trozet@redhat.com>