Adapt zrpcd to work for Carbon/Nitrogen
It updates the quagga tar.gz file which is used to
install zrpcd/quagga.
The new rpms are compiled with thrift-version 4

Zrpcd is installed as a process. Additional changes takwqe place
adopting the installation and checking of zrpcd process

Note that an extra patch is needed to update or remove the
building procedure, which is present within the project
but not used actually.

JIRA: APEX-508

Change-Id: If6da075d2aafdb8c2ecfc599f595a7201490aa23
Signed-off-by: tomsou <soth@intracom-telecom.com>