Fixes missing openstack-nova-compute on undercloud
openstack-nova-compute was missing on the undercloud image and it is
also missing in the delorean repo from which we froze artifacts for
OPNFV release.  This is likely due to the age of the repo and to fix it
we need to update the delorean repo on undercloud and install the
missing package during build time.

JIRA: APEX-538

Change-Id: I0f7cd26d5c8e4ad538b3ba3725ab5e7684a50b6a
Signed-off-by: Tim Rozet <trozet@redhat.com>